# docs/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "CompetencyX API",
        "version": "0.1.0",
        "description": "Adaptive career-roadmap backend API for assessment sessions, role inference, and next-topic recommendations."
    },
    "paths": {
        "/api/v1/accounts/me/": {
            "get": {
                "operationId": "getCurrentAccount",
                "summary": "Identify the current respondent",
                "tags": [
                    "Accounts"
                ],
                "security": [
                    {
                        "tokenAuth": []
                    },
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Account"
                                }
                            }
                        },
                        "description": "The signed-in respondent."
                    },
                    "401": {
                        "description": "No credential was presented."
                    }
                }
            }
        },
        "/api/v1/accounts/register/": {
            "post": {
                "operationId": "registerAccount",
                "summary": "Register an account",
                "tags": [
                    "Accounts"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Registration"
                            }
                        }
                    },
                    "required": true
                },
                "security": [
                    {
                        "tokenAuth": []
                    },
                    {
                        "cookieAuth": []
                    },
                    {}
                ],
                "responses": {
                    "201": {
                        "description": "Account created; the response carries the credential to hold across requests."
                    },
                    "400": {
                        "description": "Duplicate email or a password that fails validation."
                    }
                }
            }
        },
...
        "/api/v1/accounts/sign-out/": {
            "post": {
                "operationId": "signOut",
                "summary": "Sign out",
                "tags": [
                    "Accounts"
                ],
                "security": [
                    {
                        "tokenAuth": []
                    },
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "204": {
                        "description": "The credential no longer authenticates."
                    },
                    "401": {
                        "description": "No credential was presented."
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Account": {
                "type": "object",
                "description": "The identity a signed-in respondent gets back when they ask who they are.",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "readOnly": true,
                        "title": "Email address"
                    }
                },
                "required": [
                    "email",
                    "id"
                ]
            },
...
            "Registration": {
                "type": "object",
                "description": "The email and password pair a respondent presents to register or to sign in.",
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "maxLength": 150
                    },
                    "password": {
                        "type": "string",
                        "writeOnly": true
                    }
                },
                "required": [
                    "email",
                    "password"
                ]
            },
...
        },
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "sessionid"
            },
            "tokenAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "Authorization",
                "description": "Token-based authentication with required prefix \"Token\""
            }
        }
    }
}